+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.37.0] - 2022-11-12
+### Changed
+- Improved loading states for event list.
+- Swapped water week widget for an updated one with previous week results.
+
 ## [1.36.0] - 2022-10-22
 ### Changed
 - Better queries for day usage widgets.